+++ libs/sdk-core/src/lib/auth/auth.models.ts
@@ -2,3 +2,5 @@ export interface AuthTokens {
   access_token: string;
   refresh_token: string;
 }
+
+export type NucliaDBRole = 'READER' | 'WRITER' | 'MANAGER';