+++ libs/sdk-core/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add `getStandaloneKbs` method in db
 - Update `getKnowledgeBox` to also work on standalone mode
 - Add optional `uuid` in KnowledgeBox model
+- Update `getAuthHeaders` to automatically add the required `X-NUCLIADB-ROLES` depending on the method and the path
 
 # 1.1.5 (2023-03-03)